cmd/scdcentral: new package, new approach to simplify idea
The basic hosting premise is still present, but explicity more
general. Attempt to minimize definitions needed and centralize
configuration storage.

The first attempt will again be a monolithic application. But the
goal is to turn the monolithic application into a distributed
application without a re-write, only a new registry based on etcd.